+++ apps/frontend/src/pages/friends/libs/components/friends-tab/friends-tab.tsx
@@ -1,4 +1,7 @@
-import { Pagination } from "~/libs/components/components.js";
+import {
+	EmptyPagePlaceholder,
+	Pagination,
+} from "~/libs/components/components.js";
 import { EMPTY_ARRAY_LENGTH } from "~/libs/constants/constants.js";
 import { type usePagination } from "~/libs/hooks/hooks.js";
 import { type UserAuthResponseDto } from "~/modules/users/users.js";
@@ -6,28 +9,32 @@ import { type UserAuthResponseDto } from "~/modules/users/users.js";
 import { FriendList } from "../components.js";
 
 type Properties = {
+	emptyPlaceholder: string;
 	items: UserAuthResponseDto[];
 	pagination: ReturnType<typeof usePagination>;
 };
 
 const FriendsTab: React.FC<Properties> = ({
+	emptyPlaceholder,
 	items,
 	pagination,
 }: Properties) => {
 	const { handlePageChange, page, pages, pagesCount } = pagination;
 	const hasPages = items.length > EMPTY_ARRAY_LENGTH;
 
+	if (!hasPages) {
+		return <EmptyPagePlaceholder title={emptyPlaceholder} />;
+	}
+
 	return (
 		<>
 			<FriendList friends={items} />
-			{hasPages && (
-				<Pagination
-					currentPage={page}
-					onPageChange={handlePageChange}
-					pages={pages}
-					pagesCount={pagesCount}
-				/>
-			)}
+			<Pagination
+				currentPage={page}
+				onPageChange={handlePageChange}
+				pages={pages}
+				pagesCount={pagesCount}
+			/>
 		</>
 	);
 };