+++ packages/shared/src/libs/constants/constants.ts
@@ -1,2 +1,4 @@
 export { EMPTY_ARRAY_LENGTH } from "./empty-array-length.constant.js";
+export { LAST_ARRAY_ITEM } from "./last-array-item.constant.js";
 export { MAX_FILE_SIZE_IN_MB } from "./max-file-size.constant.js";
+export { PREVIOUS_INDEX_OFFSET } from "./previous-index-offset.constant.js";